Fix bug decode

+++ 19C11002_19C11003_19C11005/19C11002_19C11003_19C11005/source/EncodedImg.cpp
@@ -19,7 +19,7 @@ EncodedImg::EncodedImg(const EncodedImg& otherImg) {
 	bufferLength = otherImg.bufferLength;
 	width = otherImg.width;
 	heigth = otherImg.heigth;
-	buffer = new double[bufferLength];
+	buffer = new float[bufferLength];
 
 	if (bufferLength > 0)
 		memcpy(buffer, otherImg.buffer, bufferLength);
